# sharder: move and drain objects concurrently