--- a/pkg/apis/config/v1alpha1/defaults_test.go
+++ b/pkg/apis/config/v1alpha1/defaults_test.go
@@ -91,7 +91,8 @@ var _ = Describe("SharderConfig defaulting", func() {
 
 			Expect(obj.Controller).To(Equal(Controller{
 				Sharder: &SharderController{
-					SyncPeriod: &metav1.Duration{Duration: 5 * time.Minute},
+					SyncPeriod:      &metav1.Duration{Duration: 5 * time.Minute},
+					ConcurrentMoves: ptr.To[int32](100),
 				},
 			}))
 		})