**What this PR does / why we need it**:

Before this PR, the sharder iterated through every object of the `ControllerRing` and moved/drained them one at a time. With many objects, this becomes problematic in scenarios with dynamic controller instances. Scaling the controller deployment and replacing instances are typical scenarios. Hence, this should be handled better by the sharder.

With this PR, the sharder controller moves and drains objects concurrently (up to 100 objects by default, can be configured in `SharderConfig.controller.sharder.concurrentMoves`). This results in much faster handovers (see below), i.e., a more responsive sharded controller.

This scalability improvement becomes especially relevant if the API server is running in a different network, e.g., when using a managed Kubernetes service where you don't run control plane nodes in your cluster's network.

**Which issue(s) this PR fixes**:
Fixes https://github.com/timebertt/kubernetes-controller-sharding/issues/449

**Special notes for your reviewer**:

I compared the performance before and after this change in my test setup on STACKIT with a managed control plane (STACKIT Kubernetes Engine/Gardener). I created 1,000 `Websites` for testing different dynamic scenarios with the webhosting-operator (no changes in the operator, only in the sharder). For every scenario, I measured how long it took until all `Websites` transitioned to phase `Ready` again.

| Scenario | sequential (before) | concurrent (after) |
| --- | --- | --- |
| Delete 1 random operator pod | 3m49s | 28s |
| Scale down operator from 3 to 2 pods | 2m46s | 17s |
| Scale up operator from 2 to 3 pods | 2m06s | 17s |

The sharder first moves all "main resources" before moving "controlled resources". Accordingly, there is a short period where the main object has already been moved to a new shard, but the controlled objects are still assigned to the old shard. This causes the controller to fail (e.g., creating objects that already exist because they are not in the cache yet) until all controlled objects have been moved as well.
This has reinforcing effects in the sequential implementation, as `Websites` enter a high backoff until the sharder finishes the resync.